FirstAttempt
Attempted to use lists in order to enable multiple entries to be correct, although this did not function as intended as only the initial entry worked whilst others did not. Furthermore had trouble with adding validation checks, in stating that anything outside of the list would be deemed as incorrect.